# packages/backend/src/OpenApi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "Ballot": {
        "properties": {
          "ballot_id": {
            "type": "string"
          },
          "create_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "date_submitted": {
            "type": "number"
          },
          "election_id": {
            "type": "string"
          },
          "head": {
            "type": "boolean"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/BallotAction"
            },
            "type": "array"
          },
          "ip_hash": {
            "type": "string"
          },
          "precinct": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "update_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "user_id": {
            "type": "string"
          },
          "votes": {
            "items": {
              "$ref": "#/components/schemas/Vote"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "BallotAction": {
        "properties": {
          "action_type": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "timestamp": {
            "type": "number"
          }
        },
        "type": "object"
      },
      "Candidate": {
        "properties": {
          "bio": {
            "type": "string"
          },
          "candidate_id": {
            "type": "string"
          },
          "candidate_name": {
            "type": "string"
          },
          "candidate_url": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "party": {
            "type": "string"
          },
          "partyUrl": {
            "type": "string"
          },
          "party_url": {
            "type": "string"
          },
          "photo_filename": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Election": {
        "properties": {
          "admin_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "audit_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "auth_key": {
            "type": "string"
          },
          "claim_key_hash": {
            "type": "string"
          },
          "create_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "credential_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "election_id": {
            "type": "string"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "frontend_url": {
            "type": "string"
          },
          "head": {
            "type": "boolean"
          },
          "is_public": {
            "type": "boolean"
          },
          "owner_id": {
            "type": "string"
          },
          "races": {
            "items": {
              "$ref": "#/components/schemas/Race"
            },
            "type": "array"
          },
          "settings": {
            "$ref": "#/components/schemas/ElectionSettings"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/ElectionState"
          },
          "support_email": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "update_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "type": "object"
      },
...
      "ElectionSettings": {
        "properties": {
          "ballot_updates": {
            "type": "boolean"
          },
          "break_ties_randomly": {
            "type": "boolean"
          },
          "invitation": {
            "enum": [
              "address",
              "email"
            ],
            "type": "string"
          },
          "max_rankings": {
            "type": "number"
          },
          "public_results": {
            "type": "boolean"
          },
          "random_candidate_order": {
            "type": "boolean"
          },
          "reminders": {
            "type": "boolean"
          },
          "require_instruction_confirmation": {
            "type": "boolean"
          },
          "term_type": {
            "enum": [
              "election",
              "poll"
            ],
            "type": "string"
          },
          "time_zone": {
            "type": "string"
          },
          "voter_access": {
            "enum": [
              "closed",
              "open",
              "registration"
            ],
            "type": "string"
          },
          "voter_authentication": {
            "$ref": "#/components/schemas/authentication"
          }
        },
        "type": "object"
      },
      "ElectionState": {
        "enum": [
          "archived",
          "closed",
          "draft",
          "finalized",
          "open"
        ],
        "type": "string"
      },
...
      "NewBallot": {
        "properties": {
          "ballot_id": {
            "type": "string"
          },
          "create_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "date_submitted": {
            "type": "number"
          },
          "election_id": {
            "type": "string"
          },
          "head": {
            "type": "boolean"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/BallotAction"
            },
            "type": "array"
          },
          "ip_hash": {
            "type": "string"
          },
          "precinct": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "update_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "user_id": {
            "type": "string"
          },
          "votes": {
            "items": {
              "$ref": "#/components/schemas/Vote"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "Race": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "num_winners": {
            "type": "number"
          },
          "precincts": {
            "items": {
              "additionalProperties": false,
              "patternProperties": {
                "^[0-9]+$": {
                  "type": "string"
                }
              },
              "properties": {
                "length": {
                  "type": "number"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "race_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "voting_method": {
            "$ref": "#/components/schemas/VotingMethod"
          }
        },
        "type": "object"
      },
      "Score": {
        "properties": {
          "candidate_id": {
            "type": "string"
          },
          "score": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "Vote": {
        "properties": {
          "race_id": {
            "type": "string"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/Score"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "VotingMethod": {
        "enum": [
          "Approval",
          "IRV",
          "Plurality",
          "RankedRobin",
          "STAR",
          "STAR_PR",
          "STV"
        ],
        "type": "string"
      },
...
      "authentication": {
        "properties": {
          "address": {
            "type": "boolean"
          },
          "email": {
            "type": "boolean"
          },
          "ip_address": {
            "type": "boolean"
          },
          "phone": {
            "type": "boolean"
          },
          "registration_api_endpoint": {
            "type": "string"
          },
          "registration_data": {
            "items": [
              {
                "$ref": "#/components/schemas/registration_field"
              }
            ],
            "maxItems": 1,
            "minItems": 1,
            "type": "array"
          },
          "voter_id": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "registration_field": {
        "properties": {
          "field_name": {
            "type": "string"
          },
          "field_type": {
            "enum": [
              "photo",
              "text"
            ],
            "type": "string"
          },
          "help_text": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "paths": {
    "/Election/{id}/ballot": {
      "post": {
        "summary": "Return election ballot",
        "tags": [
          "Ballots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election ballot returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ballot": {
                      "type": "object",
                      "$ref": "#/components/schemas/NewBallot"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/ballots": {
      "get": {
        "summary": "Get ballots by election ID",
        "tags": [
          "Ballots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of ballots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    },
                    "ballots": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Ballot"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      },
      "delete": {
        "summary": "Delete all ballots for an election",
        "tags": [
          "Ballots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "All ballots deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "If the deletion was successful"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/ballot/{ballot_id}": {
      "get": {
        "summary": "Get ballot by ballot ID",
        "tags": [
          "Ballots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          },
          {
            "in": "path",
            "name": "ballot_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The ballot ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Ballot details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ballot": {
                      "type": "object",
                      "$ref": "#/components/schemas/Ballot"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ballot not found"
          }
        }
      }
    },
    "/Election/{id}/vote": {
      "post": {
        "summary": "Cast a vote in an election",
        "tags": [
          "Ballots"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Vote cast",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ballot": {
                      "type": "object",
                      "$ref": "#/components/schemas/Ballot"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
...
  },
...
}